Much cleanup / break into modules
