+++ library/Fission/Internal/Applicative.hs
@@ -1,0 +1,6 @@
+module Fission.Internal.Applicative (noop) where
+
+import RIO
+
+noop :: Applicative m => m ()
+noop = pure ()
