+++ library/Fission/Web/Auth/Client.hs
@@ -1,4 +1,4 @@
-module Fission.Web.Auth.Client (run, verify) where
+module Fission.Web.Auth.Client (verify) where
 
 import RIO
 
@@ -11,6 +11,3 @@ import qualified Fission.Web.Routes as Web
 
 verify :: BasicAuthData -> ClientM Bool
 verify = client (Proxy :: Proxy Web.AuthRoute)
-
-run :: HTTP.Manager -> BaseUrl -> ClientM a -> IO (Either ServantError a)
-run manager url query = runClientM query $ mkClientEnv manager url
